+++ requirements.txt
@@ -7,6 +7,7 @@ networkx < 2.8.3 # https://github.com/networkx/networkx/pull/5667
 numpy >= 1.22
 pydot
 pytest
+pytest-cov
 pytest-mock
 rechunker
 s3fs
